chore(front): move expectations drift types into generated api-types (#6945)
